requirements.txt
hydra-core>=1.2.0
wandb>=0.13
numpy
torch>=1.10
transformers>=4.22.0
tokenizers
requests
Flask
scikit-learn
matplotlib
langchain==0.0.225
typing_extensions==4.5
plotly
accelerate 
datasets 
rouge_score 
evaluate 
nltk 
bert_score 
wandb 
sacrebleu
